requirements.txt
bidict==0.22.0
black==22.8.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Login==0.6.2
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.1
SQLAlchemy==1.4.41
tomli==2.0.1
typing-extensions==4.3.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
